test_parameters_rt.py restored as requested
